// apps/sales-copilot/src/generated/models/setting-model.ts
export interface Setting {
  /**
   * @displayName Setting
   * @description Unique system identifier - GUID generated by system
   * @uiConstraints NEVER create input fields for this ID
   */
  id: string;
  /**
   * @displayName Setting Key
   * @description Human-readable record identifier
   * @validationRule Required for create/update operations
   */
  settingKey: string;
  /**
   * @displayName Description
   */
  description?: string;
  /**
   * @displayName Setting Value
   * @validationRule Required for create/update operations
   */
  settingValue: string;
  /**
   * @displayName Updated On
   */
  updatedOn?: string;
}

export const _Setting = 'Setting' as const;